ADDED TEST COVERAGE + LOW PULL REQUEST COMMENTS EXPLAINED
Test coverage was added in the README, as well as in the "java" directory under the "test" module. 

To explain the lack of reviews and comments from pull request #51 onwards, these pull requests were very minor changes involving the following:

1) bug fixes.
2) the addition of test cases.
3) additional documentation to make our code crystal clear.  

None of the above really required reviewers. As long as test coverage was sufficient and the test cases passed, reviewers were not needed to make sure they were correct. Additional documentation did not affect the implementation of the code. Bug fixes were a result of test cases not passing, and the discovery of minor errors that were subsequently fixed. 

Pull requests #24 to #50 are a much better representation of our process. These pull requests were made after we received your suggestions regarding increasing reviewer count, and regard code essential for proper function of our features.